Move StateFilter tests
should have done this in #14668
